backend/requirements.txt
fastapi
uvicorn[standard]
PyMuPDF
sentence-transformers
chromadb
requests
python-dotenv
python-multipart
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl


